Factor out rendering of collaborator list
